Move SyfosyketilfelleConsumer to package syfosyketilfelle
