+++ src/test/kotlin/no/nav/syfo/api/BrukerApiSpek.kt
@@ -13,7 +13,7 @@ import io.mockk.mockk
 import no.nav.syfo.api.bruker.registerBrukerApi
 import no.nav.syfo.api.bruker.urlPath39UkersVarsel
 import no.nav.syfo.consumer.PdlConsumer
-import no.nav.syfo.consumer.SyfosyketilfelleConsumer
+import no.nav.syfo.consumer.syfosyketilfelle.SyfosyketilfelleConsumer
 import no.nav.syfo.db.domain.PPlanlagtVarsel
 import no.nav.syfo.db.domain.VarselType
 import no.nav.syfo.db.storeUtsendtVarselTest
